Add incremental backups based on GMT

+++ app/lib/backup_data.py
@@ -10,10 +10,10 @@ from pathlib import Path
 import json
 import shutil
 from datetime import datetime
+import sys
 
 def get_args():
     FORMIO_URL = os.environ.get('FORMIO_URL')
-
     args = { \
             "local_host": "http://" + FORMIO_URL,
             "admin_user": "admin@example.com",
@@ -21,8 +21,27 @@ def get_args():
             "backup_dir": "backups",
             "skip_paths": ["admin"]
         }
-
     return args
+
+def update_config(dir, mode):
+    full_file = f'{dir}/version.conf'
+    result = ''
+    if mode == 'incremental':
+        if os.path.exists(full_file):
+            t = open(full_file, 'r')
+            last_incremental = t.read()
+            result = f'&modified__gt={last_incremental}'
+            t.close()
+
+        # Save with current date/time
+        current = datetime.utcnow().strftime("%Y-%m-%dT%H:%M")
+        t = open(full_file, 'w')
+        t.write(current)
+        t.close()
+    else:
+        result = ''
+    print(f'Cutoff Date: {result}')
+    return result
 
 def get_token(host, admin_user, admin_password):
     url = f'{host}/user/login'
@@ -32,7 +51,6 @@ def get_token(host, admin_user, admin_password):
             'password': admin_password
             }
         }
-
     token = ''
     try:
         r = requests.post(url, json=payload)
@@ -41,18 +59,18 @@ def get_token(host, admin_user, admin_password):
         print({ex})
     return token
 
-def get_forms(host, token, skip_paths):
+def get_forms(host, token, skip_paths, cutoff_date):
     result = []
-    url = f'{host}/form?sort=path&&limit=9999'
+    url = f'{host}/form?sort=path&&limit=9999{cutoff_date}'
     r = requests.get(url, headers={"x-jwt-token": token})
     for f in r.json():
         if not f['path'] in skip_paths:
             result.append(f)
     return result
 
-def get_submissions_for_path(path, host, token):
+def get_submissions_for_path(path, host, token, cutoff_date):
     result = []
-    url = f'{host}/{path}/submission?limit=9999'
+    url = f'{host}/{path}/submission?limit=9999{cutoff_date}'
     r = requests.get(url, headers={"x-jwt-token": token})
     for l in r.json():
         if bool(l['data']):
@@ -101,25 +119,47 @@ args = get_args()
 local_host  = args['local_host']
 local_token = get_token(local_host, args['admin_user'], args['admin_password'])
 backup_dir = args['backup_dir']
+timestamp = f'{datetime.today().strftime("%Y%m%d_%H%M")}'
 
-print(f'** Get Forms and Submissions from {local_host}')
-forms_to_pull = get_forms(local_host, local_token, args['skip_paths'])
+script_args = sys.argv[1:]
+mode = script_args[0]
+
+print(f'** {mode} backup forms and submissions from {local_host}')
+cutoff_date = update_config(backup_dir, mode)
+
+backup_started = False
+forms_to_pull = get_forms(local_host, local_token, args['skip_paths'], cutoff_date)
+print('Backup forms...')
 for form in forms_to_pull:
+    backup_started = True
     path = form['path']
     if path == 'admin':
         next
-    print(path, end='')
+    print(path)
     save_data(backup_dir, 'forms', path, 'json', form)
 
-    submissions_to_pull = get_submissions_for_path(path, local_host, local_token)
-    for submission in submissions_to_pull:
-        print('.', end='')
-        index_field = submission['_id']
-        save_data(backup_dir, path, index_field, 'json', submission)
-    print()
+print('Backup submissions...')
+submissions_to_pull = get_forms(local_host, local_token, args['skip_paths'], '')
+for form in submissions_to_pull:
+    path = form['path']
+    if path == 'admin':
+        next
 
-output_filename = f'{backup_dir}/backup_{datetime.today().strftime("%Y%m%d_%H%M")}'
-shutil.make_archive(output_filename, 'zip', f'{backup_dir}/tmp')
-shutil.rmtree(f'{backup_dir}/tmp')
+    submissions_to_pull = get_submissions_for_path(path, local_host, local_token, cutoff_date)
+    if len(submissions_to_pull) > 0:
+        print(path, end='')
+        backup_started = True
+        for submission in submissions_to_pull:
+            print('.', end='')
+            index_field = submission['_id']
+            save_data(backup_dir, path, index_field, 'json', submission)
+        print()
+
+if backup_started:
+    output_filename = f'{backup_dir}/{mode}_{timestamp}'
+    shutil.make_archive(output_filename, 'zip', f'{backup_dir}/tmp')
+    shutil.rmtree(f'{backup_dir}/tmp')
+else:
+    print ("No changes for incremental backup")
 
 print('** Done')
